requirements.txt
aiogram==3.0.0
environs~=9.0
redis
betterlogging
aiogram-dialog~=2.0.0b22
aiohttp~=3.8.5
apscheduler~=3.10.0

sqlalchemy~=2.0
alembic~=1.0
asyncpg
openai

faiss-cpu
pypdf
langchain
langchain-community
langchain-openai
langchain-core
pydantic